fix(spring-cloud-gateway): add instrument
