+++ instrument-starters/opentracing-spring-cloud-gateway-starter/src/main/java/io/opentracing/contrib/spring/cloud/gateway/GatewayTracingAutoConfiguration.java
@@ -1,5 +1,5 @@
-/*
- * Copyright 2019 The OpenTracing Authors
+/**
+ * Copyright 2017-2019 The OpenTracing Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
